backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.135.1",
    "langchain>=0.3.0",
    "langchain-openai>=0.3.0",
    "langchain-community>=0.3.0",
    "langgraph>=0.1.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
    "sse-starlette>=2.0.0",
    "pydantic>=2.0.0",
]
